Fixing insignificant, but pesky, warnings and bug in project files.
git-svn-id: http://svn.osgeo.org/metacrs/csmap/trunk@2793 4e78687f-474d-0410-85f9-8d5e500ac6b2

+++ CsMapDev/CsMapConsoleUtilities/Source/csUsefulRangeReport.cpp
@@ -340,8 +340,8 @@ long TcsCrsRange::Evaluate (TcsCrsRange& othrRange)
 	{
 		deltaWidth  = LngRange() - othrRange.LngRange ();
 		deltaHeight = LatRange() - othrRange.LatRange ();
-		widthValue   = (abs (deltaWidth)  * 100) / LngRange ();
-		heightValue  = (abs (deltaHeight) * 100) / LatRange ();
+		widthValue   = static_cast<short>((abs (deltaWidth)  * 100) / LngRange ());
+		heightValue  = static_cast<short>((abs (deltaHeight) * 100) / LatRange ());
 		if (deltaWidth  < 0) widthValue  = -widthValue;
 		if (deltaHeight < 0) heightValue = -heightValue;
 		evaluation = (widthValue + heightValue) / 2;
